spartamarket_DRF/requirements.txt
asgiref==3.8.1
autopep8==2.1.0
Django==4.2
django-cors-headers==4.3.1
django-extensions==3.2.3
django-seed==0.3.1
django-silk==5.1.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
Faker==24.11.0
gprof2dot==2022.7.29
inflection==0.5.1
packaging==24.0
pillow==10.3.0
psycopg2==2.9.9
pycodestyle==2.11.1
PyJWT==2.8.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
six==1.16.0
sqlparse==0.5.0
toposort==1.10
tzdata==2024.1
uritemplate==4.1.1
